pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bd-library-manager"
version = "0.1.0"
description = "Comic archive preparation tools - convert images, generate metadata, create CBZ archives"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "BD Library Manager contributors"}
]
keywords = ["comics", "cbz", "jxl", "comicinfo", "comic-vine"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "requests",
    "Pillow",
    "jxlpy",
    "pylibjxl",
    "numpy",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
    "mypy",
    "numpy",
]

[project.scripts]
bdlib = "bdlib.cli.main:main"

[project.entry-points."bdlib.plugins"]
general = "bdlib.plugins.general:GeneralPlugin"
converter = "bdlib.plugins.converter:ConverterPlugin"
comicvine = "bdlib.plugins.metadata.comicvine:ComicVinePlugin"

[project.urls]
Homepage = "https://github.com/yourusername/bd_library_manager"
Repository = "https://github.com/yourusername/bd_library_manager"

[tool.setuptools.packages.find]
where = ["."]
include = ["bdlib*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
